feat(arm64): Tier-3 register promotion PR-2 — full x19–x28 + leaf/float pools
Extends PR-1 (26e6f179) from 4 callee-saved regs to the complete arm64
promotion register set, gated MCC_AST_PROMOTE (default off, byte-neutral).

Pin mask -> 64-bit: ast_pinned_regs and AstScratchSave.pinned unsigned->uint64_t,
1u<< -> (uint64_t)1<< across mccgen.c / mccast.c / x86_64-gen.c. x86 is
unaffected — all its register indices are <32. This lets promo register indices
exceed 31.

Register model (arm64-gen.h/.c): MCC_NB_REGS 32->38, exposing x19–x28 at indices
28–37 (reg_classes[]=0; intr() maps them to physical x19–x28). Pools (mccast.c):
callee = {x19..x28} (10 regs, callful), caller/leaf = {x9..x15} (7 caller-saved,
no save needed), float = {v2..v5} (leaf).

Key fix — leaf/float pools miscompiled at first: they use *allocatable*
registers whose reg_classes[] carries the generic RC_INT/RC_FLOAT bit, so
gv(reg_classes[reg]) matched ANY int/float register and the entry-load of the
promoted local landed in x0/v0 instead of the target (garbage reads in the
body). Fixed by masking the generic bit at the two promo-write gv sites:
gv(reg_classes[reg] & ~(MCC_RC_INT | MCC_RC_FLOAT)) isolates the single-register
class (RC_R(x)/RC_F(x)). x86-neutral by construction — x86 promo regs are either
class-0 (load path) or single-bit (RC_R8/RC_XMM6, no generic bit), so the mask
is a no-op there (this is exactly how x86 already targets a specific register).

Leaf caller-saved is guarded against the one arm64 scalar op that lowers to a
libcall (quad long double, arm64-Linux): an arm64-gated has_call=1 when a
VT_LDOUBLE node is present routes such functions to the saved callee pool.

CMake: unified ast/replay-promote — arm64 now asserts the same set as x86
(loopy+callful+sumptr+fdot all promote).

Validated on native arm64/macOS (Mach-O): full ctest 5169/5169; default exec
296/296 byte-neutral; forced MCC_AST_PROMOTE=1 exec 296/296; exec-replay-promote
296/296; disasm confirms leaf loopy holds s/v in x9/x10 (initialized+updated)
and callful holds locals in x19–x28 across bl with stack save-restore;
self-host — mcc.c recompiled with promotion (all pools) -> a working stage2 mcc
that itself compiles and promotes correctly.

PR-3 remaining: callee-saved float pool (v8–v15) for callful float promotion;
qemu-arm64-Linux cross differential + COLOR+SPILL_SHARE self-host fixpoint on
arm64; then flip opt_promote on.

+++ src/arch/x86_64/x86_64-gen.c
@@ -2071,7 +2071,7 @@ void gen_opf(int op) {
 #if MCC_CONFIG_OPTIMIZER
 			{
 				int dr = vtop[-1].r & VT_VALMASK;
-				if (dr < VT_CONST && (ast_pinned_regs & (1u << dr))) {
+				if (dr < VT_CONST && (ast_pinned_regs & ((uint64_t)1 << dr))) {
 					int sc = get_reg(MCC_RC_FLOAT);
 					o((ft & VT_BTYPE) == VT_DOUBLE ? 0x100ff2 : 0x100ff3);
 					o(0xc0 + REG_VALUE(dr) + REG_VALUE(sc) * 8);
